refactor list_people class into module
