+++ Project/database/books.json
@@ -1,1 +1,1 @@
-[{"title":"lotr","author":"tolkien"},{"title":"lotr2","author":"tolkien"},{"title":"lotr3","author":"tolkien"},{"title":"lotr4","author":"tolkien"},{"title":"lotr5","author":"tolkien"},{"title":"lotr","author":"mimamma"}]
+[{"title":"lotr","author":"tolkien"},{"title":"lotr2","author":"tolkien"},{"title":"lotr3","author":"tolkien"},{"title":"lotr4","author":"tolkien"},{"title":"lotr5","author":"tolkien"},{"title":"lotr","author":"mimamma"},{"title":"LLama que llama","author":"zarpado"}]
